Add test for comment counter
Fix minor issues
Delete duplicate files
