+++ tests/comment.test.js
@@ -2,19 +2,20 @@
  * @jest-environment jsdom
  */
 
-import comment from '../src/comment.js';
+import comment from '../src/comment';
 import commentHTML from './_comment.html';
 
 jest.mock('../src/comment.js');
 
-beforeAll( async () => {
+beforeAll(async () => {
   document.body.innerHTML = commentHTML;
   comment.getCommenterData();
   await comment.sendComment('testID1', comment.commenterData);
+  await comment.getComments('testID1');
 });
 
-describe('', () => {
-  test('', () => {
-    expect(comment.comments).toEqual([]);
+describe('Test comment counter', () => {
+  test('Counts number of comments that were previously created', () => {
+    expect(comment.comments.length).toEqual(comment.commentCounter());
   });
 });
